Break the connection between Groups and Methods by adding another layer - the Class. A Group now contains a set of Classes, and a Class contains a set of Methods.
Still have to do something about Subclasses (I think this is going to improve the Subclass situation greatly - a Subclass will be able to live inside a Group instead of *being* a Group!)
Have to do something about the StaticClassGroup as well, since it's a group that doesn't contain any Classes (it instead contains an array of class names, as it does not contain any Methods.)

We also now error out if there are more than zero explicit groups (instead of more than one) when generating an automatic constructor.
Tested for output parity with the previous version, and everything seems to work properly.


git-svn-id: http://svn.howett.net/svn/theos/trunk@270 4410221e-0ddf-4ce3-99c0-2db6c0dbc727