Rework NIC to be more object-oriented. A file format is simply a class that one can instantiate, set variables on and build. This adds support for multiple versions of NICfiles (of which there is currently only one.)
denicify.pl has been updated to use NIC::Formats::NIC* as well. It simply sets none of the special @@ variables when building.

The theos/ symlink is (now) made by nic.pl and not by the format module.

NIC is now version 1.0. The line of dashes is generated on-the-fly so the version number can be any length ;_;.


git-svn-id: http://svn.howett.net/svn/theos/trunk@303 4410221e-0ddf-4ce3-99c0-2db6c0dbc727