[Windows] Add show in file manager for Windows Explorer.
Funnily enough, `explorer /select,filename` is the documented way to do
this, which would make you think that the filename must be part of the same
argument. However, `explorer /select, file name with spaces` works just
fine. (Cue an Old New Thing blog post explaining that this is a backwards
compatibility hack because a single program depended on it back in 1995.)