[logos] Log nil classes/selectors both with internal and Substrate.
Upside of this being implemented for Substrate is that Logos has access to
the class name as a string; Substrate only takes in a Class, which is going
to be nil if objc_getClass() failed. The class can therefore be included,
which is a huge help. Plus, HBLogError(), so it shows up in red.