[rules] Hacky fix for makedeps errors when no arch is set.
Not sure if this is the "right" fix for this or not. It seems the top make is
doing the makedeps compile, and no compilation happens there so
THEOS_CURRENT_ARCH is empty. The sub-invocations of make do have it set and
that's the only place it should be running anyway.