[deb] Remove stderr null redirect so warnings from dpkg are shown.
Custom field warnings were removed in dpkg 1.17.14.
https://github.com/guillemj/dpkg/commit/2e5f5b658c7c7153d3131e7224d9c337c0be1164

However, we still get the warning about lzma being deprecated, and there’s not
really any way to solve that until the magical day that Telesphoreo updates dpkg
for us all. (Which raises a thought that this should be target specific, not
hardcoded for all targets that might use dpkg.)