[install] Allow $THEOS_DEVICE_{PORT,USER} to be optional.
By not setting the port to 22 by default, we pass all responsibility of
calculating the port back to ssh. This is especially valuable when
$THEOS_DEVICE_IP is set to an alias defined in ~/.ssh/config which includes a
`Port` directive.

`THEOS_DEVICE_USER ?= root` remains because, well, there’s not really any way to
install packages unless you’re root. And allowing another user to sudo without a
password is just risky and stupid.