[rules] Use GitHub API to prevent no-op pulls and add messages.
No-op pulls are slow because Git is weird, so we can hit the GitHub API to see
what the latest commit is. If we get a 304 Not Modified, then there’s no new
commits.

Worth reading (even though it’s more about CocoaPods’ massive load on GitHub):
https://github.com/CocoaPods/CocoaPods/issues/4989#issuecomment-193772935

This does assume the user has curl… are there any situations where curl *won’t*
be installed on a modern Unix OS? Let alone one that Theos would be used on?
*Should* we be supporting those configurations?