[deb] Fix handling of control files without a proper file ending.
There was an edge case where if a control file did not contain an empty line it
would get the version appended on the same line as another control field,
leading to dpkg-deb throwing the error that no Version was present in the
control file. This also handles if a control file has multiple empty lines at
its end, they will be removed and a new line ending will be appended to that
last line if one is not already present.