[perl,deb] Workaround issues on WSL
Updated all relevant perls files with a proper fix that applies to only
WSL and leaves other platforms to operate normally. not that it had much
ill affect. This is just more proper. Removed on perl files that don't
need it. Add a chmod line to deb.mk to resolve a permissions issue that
occured when building packages.