[makefiles] Use "+" and pass $(MFLAGS), so we can support parallelism.
This also adds --output-sync=target to the MFLAGS, so the output doesn't look
like vomit if there are errors (although it sure makes it look less fast). This
isn't supported by (you guessed it) Apple's ancient version of make. So for now,
this requires macOS users to `brew install make` and `alias make=gmake`.