[deb] Switch from dpkg-deb to dm.pl.
Cygwin was the only platform using dm.pl; now all except iOS are using it. iOS
has dpkg-deb with the “lzma obsolete” check removed, and there are some issues
with Perl, so not switching over there. There is a dependency on a binary
component  (Compress::Raw::Lzma) so for now we’re asking users to install
IO::Compress::Lzma manually. CI updated to install that, and post-update updated
to try to install it at update-theos time.