[bundle] Distinguish “extension safe” vs “extension safe API”.
This is awkward I know. “Extension safe API” passes the app extension flag into
clang, so it knows to block you from using APIs not safe for app extensions such
as UIApplication. “Extension safe” passes the app extension flag to the linker,
so the extension safe flag will be added to the binary. Non-extension safe
binaries can be loaded into extensions anyway, but I guess it’s worth doing this
as a precaution in case that changes in future.

I kinda doubt anyone will ever need to override these anyway.