[darwin] Change blah_USE_SUBSTRATE to blah_LOGOS_DEFAULT_GENERATOR.
This is a breaking change. First, this was bad API design as it leaves no room
for other generators we may need to account for in Theos. Second, the logic is
just wrong, because the default is true, and false is an empty string, so when
we set it to false, $(or …) just skips it and falls back to our default, true.
Whoops.