[makefiles] Remove workaround for an old WSL bug.
This WSL bug was fixed about a year or two ago (the issue was ldid failing to
write to the input file in-place when the obj dir is on a Windows filesystem)
and it's causing issues right now, so just remove it.  This does mean bumping
the minimum Windows version, but I'm guessing WSL users know old versions are
buggy / lacking and they should use more recent Windows for WSL. Closes #412.