[rules] Remove -std=c99 flag; trust the compiler’s default C dialect.
Xcode no longer passes a -std=… flag unless overridden in the project
settings. Clang uses gnu11 as the current default, which, in reality,
is what probably everyone wants to/should be using now. Documentation
for clang seems to imply this default will change to newer C dialects
as time goes on, so it seems safe to rely on this behavior. Ref: #416