[_common/iphone] Add -no_data_const to fix deploying to old iOS.
iOS before 9.0 does not support __DATA_CONST. The linker now creates
binaries with __DATA_CONST, even when it should know from flags that
the binary won’t run as far back as the specified deploy version. So
we need to pass in -no_data_const to generate compatible binaries in
this case.