Added support for bundling frameworks
Frameworks go in ./Frameworks in the project root.

When initially compiled, the link paths for the frameworks in the tweak binary will be incorrect: they'll use `@rpath`, which will evaluate to the wrong path when injected into a third-party app. To correct this, we remove the `@rpath` bit of the link path then prepend the installation path of the tweak bundle.